+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Bug Fixes
 * Algebra::quadratic returns empty array when a=0 and b=0 (Previously would divide by zero)
 * Trigonometry::unitCircle division by zero error for single-point.
+* SampleData\People `getShoeSize` has a typo and was `getShowSize`. This has been fixed.
 
 ## v2.12.0 - 2025-10-17